move off_t/uint64_t guards up in API; bunch of offset math

+++ include/crc.h
@@ -22,6 +22,6 @@ static inline uint32_t crc32_final(uint32_t crc)
 
 bool crc32_validate(uint32_t expected_crc, const uint8_t header[ENTRY_HEADER_SIZE],
                     const uint8_t *key, size_t key_size,
-                    int fd, uint64_t value_pos, uint64_t value_size);
+                    int fd, off_t value_pos, size_t value_size);
 
 #endif
